Added the base of the shader system.
This frickin' subsystem has been directly responsible for the last 4 rewrites of this engine.  I think I did it right this time, but I can't be certain.  There's still a lot that needs to be done with textures and materials in order for this to be good to go, but I'll get to that in time.

A bug in the glew setup script was resolved so that compiled binaries go to the correct location, instead of the old path still included in the binary.

I tried do do a bit of documentation too, that should help later on.