Completed managed texture resources.
Textues can now be loaded, and will automatically be swapped in and out of GPU depending on the number of active users.

I think I've figured out what I'm doing with the graphics resource thing.  I'm still not sure if it's the right way to go about things, but I'm not as lost as I was before.