Updated rendering.
Fixed an order-of-operations bug that only became clear once I added multiple objects to the scene.  I was setting the uniforms after drawing the geometry, which meant that they would actually be applied to the next model rendered, instead of the current one.  Oops.

Also added a floor for testing and adjusted the static prop shader.