Added basic depth buffer psudeolinearization.
Additionally, the configuration classes were expanded to include explicit support for `float` variables.  Although `float`s and `double`s are mostly interchangeable, it is sometimes preferable to have a float pointer, instead of downcasting on dereference.

Actual buffer linearization will have to be implemented in the future, but I have already spent too much time working on this.