Added support for QOpenGLView by adding a QTSilhouetteWidget class.
An instance of the engine can now be embedded within a QT application!  That's pretty neat!

A few major changes have been made to the way the engine functions to incorporate this change.  The render loop is now no longer inherently part of the graphics subsystem.  If desired, it can be disabled by deactivating SDL, in which case an external class will have to take care of initializing OpenGL and rendering frames.

In order to accomidate changes to the graphics output system, the screen will now determine which framebuffer it should write to based on the "internal.gl.outputfbid" configuration variable.  By leaving this value as its default "0", the display will be used.  In order for QT to function, this value needs to be updated every time the screen is resized, or the buffers are regenerated.

This feature is not yet complete.  Significant additional work needs to be undertaken in order to ensure that it plays nicely with a native window (provided by SDL) and to resolve performance limiting bugs which are most likely still present.