Added Skybox.
The skybox exists as a special entity detatched from the normal fabric of reality.  All entities in the world will appear in front of the skybox, which serves specifically as a backdrop for all pixels that don't get rendered over during the standard rendering process.

It's not complete, and still needs to be adjusted to follow the camera to preserve the cubic projection's illusion.