Added entity Names.
Entities may now have names which can be used to access them in scenes when no other pointer exists to them.  I think there is still a bit of work to be done with the scene implmementation, but that's coming along nicely.

The QT test application has been completely re-worked to include a selectible entity modifier, and work has begun on adding a configuration editor and viewer.