Modified resource management system.
No longer are the days of the Graphics Resource Manager.  That silly system has been cast aside in favor of the more logical CacheableResource and LoadableResource, both of which do everything the GraphicsResource package did.  Caching and resource management is now no longer restricted to the graphics subsystem, and can be extended to apply to other systems as well.